Address PR review comments: refactor file naming utility and improve error handling
- Create shared FileHelper utility class for file name sanitization
- Remove duplicate SanitizeFileName methods from Program.cs and ScriptGenerationService.cs
- Add robust error handling to RemediatePoliciesAsync method:
  - Add null check for policy fetch results
  - Add try-catch around individual policy analysis
  - Add error handling for script generation operations
- Improve logging for better error visibility

Resolves review feedback on PR #81
