+++ ConditionalAccessExporter/Services/ScriptGenerationService.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using ConditionalAccessExporter.Models;
+using ConditionalAccessExporter.Utils;
 
 namespace ConditionalAccessExporter.Services
 {
@@ -221,42 +222,10 @@ namespace ConditionalAccessExporter.Services
                 _ => ".txt"
             };
             
-            var sanitizedPolicyName = SanitizeFileName(remediation.PolicyName);
+            var sanitizedPolicyName = FileHelper.SanitizeFileName(remediation.PolicyName);
             var fileName = $"{sanitizedPolicyName}_{format}{extension}";
             return Path.Combine(outputDirectory, fileName);
         }
-        
-        private static string SanitizeFileName(string fileName)
-        {
-            if (string.IsNullOrWhiteSpace(fileName))
-                return "unnamed";
-                
-            // Get invalid characters for file names
-            var invalidChars = Path.GetInvalidFileNameChars();
-            
-            // Replace invalid characters with underscores
-            var sanitized = new StringBuilder();
-            foreach (char c in fileName)
-            {
-                if (invalidChars.Contains(c))
-                    sanitized.Append('_');
-                else
-                    sanitized.Append(c);
-            }
-            
-            // Replace multiple consecutive underscores with single underscore
-            var result = sanitized.ToString();
-            while (result.Contains("__"))
-                result = result.Replace("__", "_");
-                
-            // Trim underscores from start and end
-            result = result.Trim('_');
-            
-            // Ensure we have a valid filename
-            if (string.IsNullOrWhiteSpace(result))
-                return "unnamed";
-                
-            return result;
-        }
+
     }
 }
